remove airplane mode from global actions dialog
I know I should explain why this is necessary. However, right now
I'm tired and in a hurry, so I'll save the explanation for later.
Probably for when I do a proper, clean, settings checkbox version of
this for CM10. (I think it would be a useful option to have in CM10,
like the checkbox that removes the screenshot button.)