Have to remove profile button from Global Actions :(
This is a stopgap measure to work around another problem that my
problematic user is having. If I can reliably reproduce the problem
then maybe I'll try to report a bug against CM 7.2. My suspicion
is that it might be something that's already fixed in CM10 and too
hard to fix in CM7.2.

(I realize I'm not describing what the problem actually is. I'm on
the road, doing this commit and about to do a rebuild in a hurry.
The explanation will have to wait, and I might not bother if I
can confirm that CM10 is not affected.)